Final Submission
This has blog module, diaper module, and bottle module. Diaper and
Article module save through. Still working on bottle module as I was
adding a modal to it and have not completed it. As for the Sockets; I
was able to successfully install them but have not integrated the
chatroom in yet.